pyproject.toml
[tool.poetry]
name = "delivery-app"
version = "0.1.0"
description = ""
authors = ["Abdunazarov <abdunazarovdior@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.2"
pydantic = {extras = ["all"], version = "^2.6.1"}
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
celery = "^5.3.6"
redis = "^5.0.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
